app/requirements.txt
fastapi[all]==0.111.1
uvicorn==0.30.3
pydantic==2.8.2
aiosqlite==0.20.0
aiohttp==3.8.2
loguru==0.7.2
asyncpg==0.26.0
google-generativeai==0.7.2